+++ src/main/java/com/newworld/saegil/facility/service/FacilityCrawlingService.java
@@ -1,11 +1,11 @@
-package com.newworld.saegil.organization.service;
+package com.newworld.saegil.facility.service;
 
 import com.newworld.saegil.location.LocationInfo;
 import com.newworld.saegil.location.LocationInfoResolveFailedException;
 import com.newworld.saegil.location.LocationInfoResolver;
-import com.newworld.saegil.organization.domain.Facility;
-import com.newworld.saegil.organization.domain.FacilityInfoSource;
-import com.newworld.saegil.organization.repository.FacilityRepository;
+import com.newworld.saegil.facility.domain.Facility;
+import com.newworld.saegil.facility.domain.FacilityInfoSource;
+import com.newworld.saegil.facility.repository.FacilityRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
