--- src/main/java/com/newworld/saegil/organization/service/FacilityCrawler.java
@@ -1,13 +1,0 @@
-package com.newworld.saegil.organization.service;
-
-import com.newworld.saegil.organization.domain.Facility;
-import com.newworld.saegil.organization.domain.FacilityInfoSource;
-
-import java.util.List;
-
-public interface FacilityCrawler {
-
-    FacilityInfoSource getSupportingFacilityInfoSource();
-
-    List<Facility> crawl();
-}
